fixes #10. dont ping the same IP address from multiple goroutines
ICMP doesnt use ports, so when waiting for PING replies only the
src address is listened for. If we have two processes sending pings,
then both processes will recieve the pings, which confuses things.
This commit addresses this by using a co-ordinator to allow concurrent
ping requests for the same IP to share the results.